requirements.txt
PyOpenGL
PyQt5
matplotlib
numpy
scipy
pygame
imageio
plyfile
scikit-image